requirements.txt
pytesseract==0.3.10
Pillow==10.4.0
pdfplumber==0.11.4
pdf2image==1.17.0
rapidfuzz==3.9.6
pandas==2.2.2
jinja2==3.1.4
beautifulsoup4==4.12.3
numpy==2.0.2


